# API/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Citi+
  description: Documentation for Citi+ REST API.
  termsOfService: https://www.ratp.fr/gestion-de-vos-donnees-personnelles
  contact:
    email: l_laugier@etu-webschoolfactory.fr
  license:
    name: MIT License
    url: http://localhost:80/api/v1/documentation/LICENSE
  version: 1.0.0
...
servers:
- url: http://localhost:80/api/v1/
...
paths:
  /zones-qualite-air:
    get:
      tags:
      - airparif
      summary: Get a GeoJSON feature collection of chosen areas with air quality data
        as properties
      operationId: ZonesGET
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/featureCollection'
        500:
          description: Database error
          content: {}
  /indices-qualite-air/{arrondissement}:
    get:
      tags:
      - airparif
      summary: Get air quality data by arrondissement number for the last 7 days
      description: Use "all" for an average of all arrondissements.
      operationId: DataGET
      parameters:
      - name: arrondissement
        in: path
        description: 'example: 15'
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/airQuality'
        500:
          description: Database error
          content: {}
  /arrondissements-qualite-air:
    get:
      tags:
      - airparif
      summary: Get a GeoJSON feature collection of areas in each arrondissement with
        air quality data as properties
      operationId: ArrondissementsGET
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/featureCollection'
        500:
          description: Database error
          content: {}
...
  /tables-bdd:
    get:
      tags:
      - database
      summary: Get all table names from database
      operationId: TablesGET
      parameters:
      - name: limit
        in: query
        schema:
          type: string
      - name: offset
        in: query
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/table'
        500:
          description: Database error
          content: {}
  /colonnes-table/{table}:
    get:
      tags:
      - database
      summary: Get column names by table name
      operationId: ColsGET
      parameters:
      - name: table
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cols'
        500:
          description: Database error
          content: {}
  /rangees-table/{table}:
    get:
      tags:
      - database
      summary: Get string records by table name
      operationId: RowsGET
      parameters:
      - name: table
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rows'
        500:
          description: Database error
          content: {}
...
components:
  schemas:
    airQuality:
      type: object
      properties:
        O3:
          type: array
          items:
            type: number
            format: double
        NO2:
          type: array
          items:
            type: number
            format: double
        PM10:
          type: array
          items:
            type: number
            format: double
        PM25:
          type: array
          items:
            type: number
            format: double
    featureCollection:
      type: object
      properties:
        type:
          type: string
          example: FeatureCollection
        features:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                example: Feature
              geometry:
                  type: object
                  properties:
                    type:
                      type: string
                      example: Point
                    coordinates:
                      type: array
                      items:
                        type: array
                        items:
                            type: number
                            format: double
                  required:
                    - type
                    - coordinates
            required:
            - geometry
...
    table:
      type: array
      items:
        type: string
        example: trajets
    cols:
      type: array
      items:
        type: string
        example: id_user
    rows:
      type: array
      items:
        type: string
        example: "80"
...